pyproject.toml
[project]
name = "appaltigpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastembed>=0.7.4",
    "ipython>=8.37.0",
    "llama-index>=0.14.10",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "marimo>=0.18.4",
    "mistralai>=1.10.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
